Fix parse error if a TYPED_ATTRS field is not found
Previously, a parse error would occur if the changeset XML was missing
any of the fields `id`, `uid`, `min_lon`, `min_lat`, `max_lon`,
`max_lat`, `comments_count`, `changes_count`, or `open`. This commit
changes the behavior so that instead, if a field is missing from the
input XML, it will also be missing from the output object. This allows
callers to decide how to handle the missing field.